ant_test now works
Note that the values still don't mean much - the ant_gains are
arbitrary, etc. But we can start to use this to ensure the linear
algebra is correct by putting in specific data into the ant_gain file
and setting up test cases.

+++ Mymath.h
@@ -3,9 +3,9 @@
 
 #define PI 3.141592654
 
-float cosDegrees(float degrees);
-float sinDegrees(float degrees); 
-float tanDegrees(float degrees); 
+float cosDegrees(int degrees);
+float sinDegrees(int degrees); 
+float tanDegrees(int degrees); 
 float acosDegrees(float value); 
 float asinDegrees(float value); 
 float atanDegrees(float value);
